Fix for setting NULL
In different cases if you set NULL to something, then it is checked
with `isset()` in changed, content and other arrays when read.

But in PHP

``` php

$arr = array('key' => NULL);
isset($arr['key']); // FALSE, becase
isset(NULL); // FALSE
```

So if you set NULL to association, then it does not count it as changed.