Fix find_by_slug return value
When there are no results, the call to the `first()` method returns
NULL, but this does not get the event system to return the NULL value.

This change makes it so `find_by_slug` will return `FALSE` when there
are no results.

If you have previously checked explicitly for the NULL value, you had
been affected by this bug. This commit does not change that, you are
still affected and you to check for falsy value. Changing the event
system to be able to return NULL instead of chaining and returning the
builder object is not an easy thing and will lead to more BC changes.

If you have previously checked for falsy return value from find_by_slug,
you have been affected by this bug and you will no longer be affected
with this commit.