Fix triggering callbacks with upperase letters
Jam is lower-casing the names of the methods when discovering events.
So when a method is called later through the `__call()` and `trigger_callback()`
methods, the method name should be again lower-cased so it could be find in the events map.

This is fixing registering `model_call_` methods in a behavior with camel case.